split DirectoryListItem between artistas and miembros. move Artistas and Miembros componenents into directories.
